Always explicitly create directories
Before creating nested directories with "mkdir -p", we should always
create the $LE_WORKING_DIR and $LE_CONFIG_HOME directories, if they do
not exist. This will ensure, that the directories will get the
apropriate permission when created.

If the user created the directories before the execution or modified the
permissions between executions, the permissions of the directories will
not be modified.